+++ iLand/World/CarbonNitrogenTuple.cs
@@ -8,34 +8,34 @@
     /// </remarks>
     public class CarbonNitrogenTuple
     {
-        public double C { get; set; } // carbon pool
-        public double N { get; set; } // nitrogen pool
+        public float C { get; set; } // carbon pool
+        public float N { get; set; } // nitrogen pool
 
         public CarbonNitrogenTuple()
         {
-            C = 0.0;
-            N = 0.0;
+            C = 0.0F;
+            N = 0.0F;
         }
 
-        public CarbonNitrogenTuple(double carbonTonsHa, double nitrogenTonsHa) 
+        public CarbonNitrogenTuple(float carbonTonsHa, float nitrogenTonsHa) 
         { 
             C = carbonTonsHa; 
             N = nitrogenTonsHa; 
         }
 
-        public bool IsEmpty() { return C == 0.0; } // returns true if pool is empty
-        public bool IsValid() { return C >= 0.0 && N >= 0.0; } // return true if pool is valid (content of C or N >=0)
-        public double CNratio() { return N > 0 ? C / N : 0.0; } // current CN ratio
+        public bool IsEmpty() { return C == 0.0F; } // returns true if pool is empty
+        public bool IsValid() { return C >= 0.0F && N >= 0.0F; } // return true if pool is valid (content of C or N >=0)
+        public float CNratio() { return N > 0.0F ? C / N : 0.0F; } // current CN ratio
                                                            /// retrieve the amount of biomass (kg/ha). Uses the global C-fraciton. Soil pools are in t/ha!!!
-        public double Biomass() { return C / Constant.BiomassCFraction; }
+        public float Biomass() { return C / Constant.BiomassCFraction; }
         // some simple operators
         public static CarbonNitrogenTuple operator +(CarbonNitrogenTuple p1, CarbonNitrogenTuple p2) { return new CarbonNitrogenTuple(p1.C + p2.C, p1.N + p2.N); } // return the sum of two pools
         public static CarbonNitrogenTuple operator -(CarbonNitrogenTuple p1, CarbonNitrogenTuple p2) { return new CarbonNitrogenTuple(p1.C - p2.C, p1.N - p2.N); } // return the difference of two pools
-        public static CarbonNitrogenTuple operator *(CarbonNitrogenTuple p, double factor) { return new CarbonNitrogenTuple(p.C * factor, p.N * factor); } // return the pool multiplied with 'factor'
+        public static CarbonNitrogenTuple operator *(CarbonNitrogenTuple p, float factor) { return new CarbonNitrogenTuple(p.C * factor, p.N * factor); } // return the pool multiplied with 'factor'
 
         /// add biomass to the pool (kg dry mass/ha); CNratio is used to calculate the N-Content, the global C-Fraction of biomass is used to
         /// calculate the amount of carbon of 'biomass'.
-        public void AddBiomass(double biomass, double cnRatio)
+        public void AddBiomass(float biomass, float cnRatio)
         { 
             this.C += Constant.BiomassCFraction * biomass; 
             this.N += Constant.BiomassCFraction * biomass / cnRatio;
@@ -43,8 +43,8 @@
 
         public void Clear() 
         {
-            C = 0.0; 
-            N = 0.0; 
+            C = 0.0F; 
+            N = 0.0F; 
         }
     }
 }
